change salary type

+++ Models/EmployeeDetails.cs
@@ -4,7 +4,7 @@
     {
         public string FirstName { get; set; }
         public string LastName { get; set; }
-        public decimal AnnualSalary { get; set; }
+        public uint AnnualSalary { get; set; }
         private decimal superRate;
         public decimal SuperRate
         {
